Introduce StreamResponseBase
